fix: <html lang>をroot layout統合によりSSR出力する(Issue #293)
src/app/[locale]/layout.tsxをroot layout化し、<html lang={locale}>を初期HTMLに
直接出力する。従来のbeforeInteractiveスクリプトによる後付け(document.documentElement.lang)は
削除。src/app/layout.tsxは廃止し、src/app/直下に残るnot-found.tsx/error.tsxは
上位layoutを持たないため自前で<html>/<body>を提供する自己完結ファイルに変更
(global-error.tsxと同じパターン)。ルートpage.tsx(常にmiddlewareに先取りされる
"/"リダイレクト)はroot layoutを必要としないroute.tsに置き換えた。
