--- src/app/page.tsx
@@ -1,7 +1,0 @@
-import { redirect } from "next/navigation"
-
-const Page = () => {
-  // Root page will be handled by middleware to redirect to default locale
-  redirect("/ja")
-}
-export default Page
